adapted monitoring example
