--- samples/monitor/E3_ExtractURLs/function.json
@@ -1,10 +1,0 @@
-{
-  "scriptFile": "__init__.py",
-  "bindings": [
-    {
-      "name": "issues",
-      "type": "activityTrigger",
-      "direction": "in"
-    }
-  ]
-}
