+++ samples/monitor/E3_Monitor/function.json
