+++ samples/monitor/E3_TooManyOpenIssues/function.json
