+++ samples/monitor/local.settings.json
@@ -2,6 +2,9 @@
   "IsEncrypted": false,
   "Values": {
     "AzureWebJobsStorage": "",
-    "FUNCTIONS_WORKER_RUNTIME": "python"
+    "FUNCTIONS_WORKER_RUNTIME": "python",
+    "TwilioPhoneNumber": "",
+    "TwilioAccountSid": "",
+    "TwilioAuthToken": ""
   }
 }
